Scraper v.2
The scraper uses the urllib3 library. Most guides and tutorials are for urllib 2. This scraper is built to get stock prices. I will eventually add functionality to add to a speadsheet but in the meantime, this code actually does not work because it triggers the "Are you a robot page" so none of the figures that the program looks for are there. I am currently looking for a work around.