Create Scraper v.3
After much trial and error, I've finally figured out that in Python 3 using urllib3, you need to have headers that consist of at least 'User-Agent' and 'authority' (at least for the Bloomberg website). This will allow you to get past the "Are you a robot page" and avoid HTTP error 403.